+++ CHANGES.rst
@@ -12,6 +12,7 @@ Bug Fixes
 - ``pip`` dependencies in ``environment.yml`` files for the ``mamba`` plugin are handled correctly (#1326)
 - ``asv.config.json`` matrix requirements no longer need ``pip+`` set explicitly
   for calling the ``pip`` solver for ``virtualenv``
+- ``asv`` will now use ``conda_environment_file`` if it exists (#1325)
 
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^